TaskStrip/requirements.txt
Django==4.1.3
python-dotenv==0.21.0
djangorestframework==3.14.0
stripe==5.0.0
